Create movement_control.sv
it will take in an input command to tell it to find the person, this command makes the bot rotate 45 degrees at its current position, then pauses for 0.5 seconds, then rotates again. This will happen 8 times until the robot is rotated by to its original orientation. at every pause, it will output a flag to tell the pitch detect module to record the amplitude. Once this action is complete, it will output a finish flag to programmer module to tell that the rotation is finished. this module also takes in an orientation command to tell it to rotate by a certain degrees to face that direction.