Create programmer.sv
this module take in the finish rotate flag from the movement module, and a finish recording flag from pitch detect module. When these two flags are all high, it will use the output of pitch detect module, an array of data, this data contains the corresponding recorded highest amplitude when facing at each direction at each pause. Here we use the amplitude to calculate where the noise is coming from, therefore we have to calculate which direction with the provided amplitude, then the corresponding direction will be output to movement module for it to turn the bot to face that direction